services/api/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
python-multipart>=0.0.18
python-dotenv>=1.0.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
boto3>=1.35.0
python-magic>=0.4.27
Pillow>=11.0.0
PyPDF2>=3.0.0
